requirements.txt
aiofiles==0.5.0
aiomysql==0.0.20
aiosqlite==0.13.0
aniso8601==7.0.0
apispec==3.3.0
apistar==0.7.2
asyncpg==0.20.1
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
ciso8601==2.1.3
click==7.1.2
cryptography==2.9.2
docopt==0.6.2
graphene==2.1.8
graphql-core==2.3.1
graphql-relay==2.0.1
graphql-server-core==2.0.0
h11==0.9.0
httptools==0.0.13
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
marshmallow==3.6.0
numpy==1.18.4
promise==2.3
pycparser==2.20
PyMySQL==0.9.2
PyPika==0.37.6
python-multipart==0.0.5
PyYAML==5.3.1
requests==2.23.0
requests-toolbelt==0.9.1
responder==2.0.5
rfc3986==1.4.0
Rx==1.6.1
six==1.14.0
starlette==0.12.13
tortoise-orm==0.16.11
typesystem==0.2.4
typing-extensions==3.7.4.2
urllib3==1.25.9
uvicorn==0.10.9
uvloop==0.14.0
websockets==8.1
whitenoise==5.0.1
